Firebase GoogleAuth based on GoogleSignIn Sample
All users use anonymous authentication if they don't login but in order
to access data with other devices(UID is device specific for anonAuth)
the user has to login with Google. When loging in with Google, the anon
and Google accounts are linked, keeping the same UID so no changes in
the database are required. If the user logs out of Google, the account
is downgraded back to anon(TODO: move data to new anon account UID)

To determine: how to move data from deleted Google account to downgraded
account.