Fix invalid Google credentials, add firebase functions to delete user and data
Before any major change to the Firebase user authenticated with
Google(change email, password, etc., in our case signout), the user must
be reauthenticated with GoogleAuth credentials so we now ensure the user
is reauthenticated when entering the GoogleSigInActivity.

Implement a firebase functions that catches an app_remove event log from
analytics to delete the user and the user data if he is anonymous. Since
the app is uninstalled we loose access to the user and it's impossible
to determine his userUID so the data is left useless in the database.
Add node_modules to .gitignore