Many, many changes...
Criteria now keeps a list of items
Each item now has a list of filenames, each filename has a dbkey(for
firebase) and a deleted flag
Items deleted are moved to a trash folder, and the item and it's files
are flagged deleted. Items can be restore in the settings, or
permanently deleted. (TODO: clean trash and delete specific file in
item)
Fix Dropbox and MEOCloud not being initialized in AddItemActivity if app
was closed
New tasks to create folder for MEOCloud and move file in Dropbox and
MEOCloud.
Fix lists not being updated when new data is received(not properly
working, the list is being drawn everytime a new item is
added/edited/deleted from firebase calls and the updateList with
notifyDataSetChanged doesn't work...)
MEOCallback onComplete now returns the Object and not the full response.
RemotePath in util package with useful method to get file paths for
MEOCloud and Dropbox