Several fixed bugs and additions
Move files if the criteria was changed when editing item
If the user is online:

If the user has connected services then files within the local
directories should be uploaded accordingly with the same dir structure -
INCOMPLETE - we currently don't have a way to detect if all services
upload the file successfuly so the local file is always deleted if he
has internet.
If the destination file already exists in MEO/Drop then delete the
source file(in case the file was moved to the trash while offline, in
the cloud the file remains in the same position, so when restoring the
file online it'll delete from the trash)

Fix user name and email not being shown in the MainActivity.
Profile pic in NavigationDrawer now has fixed size
Update the local files according to the actions in the Lists(rename,
delete, restore, permanent delete)